# Memory consumption improvements (less append-reallocs)