As I mentioned in another [PR](https://github.com/valyala/fastjson/pull/101) when processing large files we have a large number of memory reallocations associated with reaching capacity for a slice.
In this PR I modified the benchmarks so that they give us a more realistic estimate of memory, relying only on the data generated in a particular test (got rid of shared pool).

I achieved a reduction in memory consumption by getting rid of slice reallocations altogether by adding a chain of linked lists

Now for the proof of performance.

Here are the results of the tests
```
go test --run=None --bench="BenchmarkParse/.*/fastjson" --memprofile=mem.out
```

BEFORE
```
BenchmarkParse/small/fastjson-8       	41621682	        29.88 ns/op	6358.96 MB/s	       0 B/op	       0 allocs/op
BenchmarkParse/small/fastjson-get-8   	24688966	        47.32 ns/op	4015.26 MB/s	       0 B/op	       0 allocs/op
BenchmarkParse/medium/fastjson-8      	 4803705	       247.9 ns/op	9395.82 MB/s	       0 B/op	       0 allocs/op
BenchmarkParse/medium/fastjson-get-8  	 4600939	       260.5 ns/op	8939.31 MB/s	       0 B/op	       0 allocs/op
BenchmarkParse/large/fastjson-8       	  357400	      3314 ns/op	8484.08 MB/s	       8 B/op	       0 allocs/op
BenchmarkParse/large/fastjson-get-8   	  359676	      3276 ns/op	8582.07 MB/s	       8 B/op	       0 allocs/op
BenchmarkParse/canada/fastjson-8      	    1044	   1203621 ns/op	1870.24 MB/s	  404343 B/op	    2489 allocs/op
BenchmarkParse/canada/fastjson-get-8  	    1074	   1092411 ns/op	2060.63 MB/s	  393048 B/op	    2419 allocs/op
BenchmarkParse/citm/fastjson-8        	    3796	    316964 ns/op	5449.21 MB/s	   34197 B/op	     191 allocs/op
BenchmarkParse/citm/fastjson-get-8    	    3456	    342819 ns/op	5038.24 MB/s	   37561 B/op	     209 allocs/op
BenchmarkParse/twitter/fastjson-8     	   17029	     80574 ns/op	7837.72 MB/s	    2804 B/op	       6 allocs/op
BenchmarkParse/twitter/fastjson-get-8 	   15583	     73254 ns/op	8620.89 MB/s	    3064 B/op	       7 allocs/op
BenchmarkParse/20mb/fastjson-8        	     109	   9272379 ns/op	2209.87 MB/s	22814671 B/op	   89938 allocs/op
BenchmarkParse/20mb/fastjson-get-8    	     116	   8675982 ns/op	2361.78 MB/s	21437947 B/op	   84510 allocs/op
```
![image](https://github.com/user-attachments/assets/9b8170a8-e2b7-423d-84cf-d08d915afec7)

AFTER
```
BenchmarkParse/small/fastjson-8       	36022194	        34.03 ns/op	5583.94 MB/s	       0 B/op	       0 allocs/op
BenchmarkParse/small/fastjson-get-8   	22143430	        52.25 ns/op	3636.07 MB/s	       0 B/op	       0 allocs/op
BenchmarkParse/medium/fastjson-8      	 4418314	       270.3 ns/op	8615.86 MB/s	       0 B/op	       0 allocs/op
BenchmarkParse/medium/fastjson-get-8  	 4103462	       292.2 ns/op	7970.02 MB/s	       0 B/op	       0 allocs/op
BenchmarkParse/large/fastjson-8       	  325362	      3596 ns/op	7818.30 MB/s	       5 B/op	       0 allocs/op
BenchmarkParse/large/fastjson-get-8   	  314974	      3658 ns/op	7686.62 MB/s	       5 B/op	       0 allocs/op
BenchmarkParse/canada/fastjson-8      	     925	   1270325 ns/op	1772.03 MB/s	  194292 B/op	    2320 allocs/op
BenchmarkParse/canada/fastjson-get-8  	     838	   1293470 ns/op	1740.33 MB/s	  214463 B/op	    2561 allocs/op
BenchmarkParse/citm/fastjson-8        	    2630	    381243 ns/op	4530.45 MB/s	   22446 B/op	     201 allocs/op
BenchmarkParse/citm/fastjson-get-8    	    2632	    383046 ns/op	4509.14 MB/s	   22429 B/op	     201 allocs/op
BenchmarkParse/twitter/fastjson-8     	   15778	     76047 ns/op	8304.27 MB/s	    1640 B/op	       5 allocs/op
BenchmarkParse/twitter/fastjson-get-8 	   15397	     76405 ns/op	8265.33 MB/s	    1681 B/op	       6 allocs/op
BenchmarkParse/20mb/fastjson-8        	     138	   8586573 ns/op	2386.37 MB/s	 9507728 B/op	   60292 allocs/op
BenchmarkParse/20mb/fastjson-get-8    	     134	   8658302 ns/op	2366.60 MB/s	 9791524 B/op	   62092 allocs/op
```

![image](https://github.com/user-attachments/assets/14cff6ca-f3eb-4bca-8227-b9c0be4a734b)


As you can see, the number of memory allocations hasn't changed much, but the overall memory consumption (B/ops) has decreased by almost three times. This is especially evident on tests with huge files.

I added a 20 megabyte file to the tests to achieve dramatic effect