@StarpTech 2024-11-11
Hi @iv-menshenin, thanks for the PR. Could you attach a report with https://pkg.go.dev/golang.org/x/perf/cmd/benchstat that gives us a report of before and after. What I immediately recognize is a significant drop in performance for smaller datasets. If this can't be fixed, it is unlikely that we will merge it.