@iv-menshenin 2024-11-11
I admit my mistake, the only thing I did was to push the changes from fastjson to your repo and didn't notice some difference.
For example this
```
p.c = &cache{vs: make([]Value, 4)}
```
Also, I didn't want to re-immerse myself in the problem or do extensive testing. However, your request to attach benchstat prompted me to take action.

After analyzing the results (completely unsatisfying), I realized that my case was very narrow and made the following refinements:
 - I took into account the peculiarities of your repo
- improved memory handling of not only arrays, but also objects with a large number of keys-values