@iv-menshenin 2024-11-11
```
pkg: github.com/wundergraph/astjson
cpu: Intel(R) Core(TM) i7-9700F CPU @ 3.00GHz
                  │ BenchmarkArenaTypicalUse-before.txt │ BenchmarkArenaTypicalUse-after.txt │
                  │               sec/op                │   sec/op     vs base               │
ArenaTypicalUse-8                           42.32n ± 0%   43.65n ± 0%  +3.14% (p=0.000 n=10)

                  │ BenchmarkArenaTypicalUse-before.txt │ BenchmarkArenaTypicalUse-after.txt  │
                  │                 B/s                 │     B/s       vs base               │
ArenaTypicalUse-8                          2.729Gi ± 0%   2.646Gi ± 0%  -3.05% (p=0.000 n=10)

                  │ BenchmarkArenaTypicalUse-before.txt │ BenchmarkArenaTypicalUse-after.txt │
                  │                B/op                 │      B/op       vs base            │
ArenaTypicalUse-8                            0.000 ± 0%       0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                  │ BenchmarkArenaTypicalUse-before.txt │ BenchmarkArenaTypicalUse-after.txt │
                  │              allocs/op              │   allocs/op     vs base            │
ArenaTypicalUse-8                            0.000 ± 0%       0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

```
pkg: github.com/wundergraph/astjson
cpu: Intel(R) Core(TM) i7-9700F CPU @ 3.00GHz
                                            │ BenchmarkParseRawString-before.txt │  BenchmarkParseRawString-after.txt   │
                                            │               sec/op               │    sec/op     vs base                │
ParseRawString/""-8                                                0.6833n ±  1%   0.6833n ± 0%        ~ (p=0.955 n=10)
ParseRawString/"a"-8                                               0.8707n ± 17%   0.7368n ± 1%  -15.37% (p=0.023 n=10)
ParseRawString/"abcd"-8                                            0.8715n ±  0%   0.8799n ± 2%        ~ (p=0.085 n=10)
ParseRawString/"abcdefghijk"-8                                     0.8993n ±  1%   0.8816n ± 1%   -1.97% (p=0.000 n=10)
ParseRawString/"qwertyuiopasdfghjklzxcvb"-8                         1.022n ±  1%    1.050n ± 1%   +2.69% (p=0.000 n=10)
geomean                                                            0.8622n         0.8366n        -2.97%

                                            │ BenchmarkParseRawString-before.txt │  BenchmarkParseRawString-after.txt   │
                                            │                B/s                 │     B/s       vs base                │
ParseRawString/""-8                                                2.726Gi ±  1%   2.726Gi ± 0%        ~ (p=0.971 n=10)
ParseRawString/"a"-8                                               3.209Gi ± 20%   3.792Gi ± 1%  +18.15% (p=0.023 n=10)
ParseRawString/"abcd"-8                                            6.412Gi ±  0%   6.351Gi ± 2%        ~ (p=0.089 n=10)
ParseRawString/"abcdefghijk"-8                                     13.46Gi ±  1%   13.73Gi ± 1%   +2.01% (p=0.000 n=10)
ParseRawString/"qwertyuiopasdfghjklzxcvb"-8                        23.69Gi ±  1%   23.08Gi ± 1%   -2.60% (p=0.000 n=10)
geomean                                                            7.088Gi         7.305Gi        +3.06%

                                            │ BenchmarkParseRawString-before.txt │  BenchmarkParseRawString-after.txt  │
                                            │                B/op                │    B/op     vs base                 │
ParseRawString/""-8                                                 0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawString/"a"-8                                                0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawString/"abcd"-8                                             0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawString/"abcdefghijk"-8                                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawString/"qwertyuiopasdfghjklzxcvb"-8                         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                        ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                            │ BenchmarkParseRawString-before.txt │  BenchmarkParseRawString-after.txt  │
                                            │             allocs/op              │ allocs/op   vs base                 │
ParseRawString/""-8                                                 0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawString/"a"-8                                                0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawString/"abcd"-8                                             0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawString/"abcdefghijk"-8                                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawString/"qwertyuiopasdfghjklzxcvb"-8                         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                        ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

```
pkg: github.com/wundergraph/astjson
cpu: Intel(R) Core(TM) i7-9700F CPU @ 3.00GHz
                                    │ BenchmarkParseRawNumber-before.txt │   BenchmarkParseRawNumber-after.txt   │
                                    │               sec/op               │    sec/op      vs base                │
ParseRawNumber/1-8                                         0.4634n ±  1%   0.4396n ±  0%   -5.11% (p=0.000 n=10)
ParseRawNumber/1234-8                                      0.6455n ± 11%   0.5808n ±  2%  -10.02% (p=0.000 n=10)
ParseRawNumber/123456-8                                    0.8386n ±  1%   0.6812n ±  0%  -18.76% (p=0.000 n=10)
ParseRawNumber/-1234-8                                     0.7809n ±  0%   0.6928n ± 10%  -11.27% (p=0.000 n=10)
ParseRawNumber/1234567890.1234567-8                         1.479n ±  1%    1.486n ±  1%        ~ (p=0.072 n=10)
ParseRawNumber/-1.32434e+12-8                               1.326n ±  0%    1.311n ±  1%   -1.13% (p=0.000 n=10)
geomean                                                    0.8526n         0.7854n         -7.88%

                                    │ BenchmarkParseRawNumber-before.txt │   BenchmarkParseRawNumber-after.txt   │
                                    │                B/s                 │      B/s       vs base                │
ParseRawNumber/1-8                                         2.010Gi ±  1%   2.118Gi ±  0%   +5.38% (p=0.000 n=10)
ParseRawNumber/1234-8                                      5.792Gi ± 10%   6.414Gi ±  2%  +10.76% (p=0.000 n=10)
ParseRawNumber/123456-8                                    6.664Gi ±  1%   8.203Gi ±  0%  +23.10% (p=0.000 n=10)
ParseRawNumber/-1234-8                                     5.964Gi ±  0%   6.747Gi ± 10%  +13.14% (p=0.000 n=10)
ParseRawNumber/1234567890.1234567-8                        11.33Gi ±  1%   11.28Gi ±  0%        ~ (p=0.075 n=10)
ParseRawNumber/-1.32434e+12-8                              8.427Gi ±  0%   8.523Gi ±  1%   +1.14% (p=0.000 n=10)
geomean                                                    5.946Gi         6.455Gi         +8.56%

                                    │ BenchmarkParseRawNumber-before.txt │  BenchmarkParseRawNumber-after.txt  │
                                    │                B/op                │    B/op     vs base                 │
ParseRawNumber/1-8                                          0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawNumber/1234-8                                       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawNumber/123456-8                                     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawNumber/-1234-8                                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawNumber/1234567890.1234567-8                         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawNumber/-1.32434e+12-8                               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                    │ BenchmarkParseRawNumber-before.txt │  BenchmarkParseRawNumber-after.txt  │
                                    │             allocs/op              │ allocs/op   vs base                 │
ParseRawNumber/1-8                                          0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawNumber/1234-8                                       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawNumber/123456-8                                     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawNumber/-1234-8                                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawNumber/1234567890.1234567-8                         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ParseRawNumber/-1.32434e+12-8                               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

```

```
pkg: github.com/wundergraph/astjson
cpu: Intel(R) Core(TM) i7-9700F CPU @ 3.00GHz
                                    │ BenchmarkObjectGet-before.txt │     BenchmarkObjectGet-after.txt      │
                                    │            sec/op             │    sec/op      vs base                │
ObjectGet/items_10/lookups_0-8                         33.86n ±  3%    39.98n ±  3%  +18.09% (p=0.000 n=10)
ObjectGet/items_10/lookups_1-8                         47.55n ±  2%    54.93n ±  3%  +15.51% (p=0.000 n=10)
ObjectGet/items_10/lookups_2-8                         52.49n ±  2%    60.00n ±  0%  +14.30% (p=0.000 n=10)
ObjectGet/items_10/lookups_4-8                         62.63n ±  2%    70.30n ±  2%  +12.25% (p=0.000 n=10)
ObjectGet/items_10/lookups_8-8                         80.18n ±  2%    87.34n ±  2%   +8.92% (p=0.000 n=10)
ObjectGet/items_10/lookups_16-8                        115.8n ±  1%    124.9n ±  1%   +7.90% (p=0.000 n=10)
ObjectGet/items_10/lookups_32-8                        186.6n ±  0%    197.7n ±  1%   +5.95% (p=0.000 n=10)
ObjectGet/items_10/lookups_64-8                        328.4n ±  1%    345.4n ±  0%   +5.18% (p=0.000 n=10)
ObjectGet/items_100/lookups_0-8                        384.5n ±  0%    428.9n ±  1%  +11.53% (p=0.000 n=10)
ObjectGet/items_100/lookups_1-8                        411.8n ±  1%    460.1n ±  1%  +11.74% (p=0.000 n=10)
ObjectGet/items_100/lookups_2-8                        437.9n ±  0%    488.9n ±  0%  +11.65% (p=0.000 n=10)
ObjectGet/items_100/lookups_4-8                        488.7n ±  0%    543.6n ±  0%  +11.24% (p=0.000 n=10)
ObjectGet/items_100/lookups_8-8                        588.0n ±  1%    646.8n ±  0%  +10.01% (p=0.000 n=10)
ObjectGet/items_100/lookups_16-8                       786.4n ±  0%    857.1n ±  0%   +9.00% (p=0.000 n=10)
ObjectGet/items_100/lookups_32-8                       1.185µ ±  1%    1.271µ ±  0%   +7.26% (p=0.000 n=10)
ObjectGet/items_100/lookups_64-8                       1.976µ ±  0%    2.116µ ±  0%   +7.09% (p=0.000 n=10)
ObjectGet/items_1000/lookups_0-8                       3.947µ ±  0%    4.394µ ±  0%  +11.34% (p=0.000 n=10)
ObjectGet/items_1000/lookups_1-8                       4.178µ ±  0%    4.618µ ±  0%  +10.53% (p=0.000 n=10)
ObjectGet/items_1000/lookups_2-8                       4.401µ ±  0%    4.864µ ±  1%  +10.53% (p=0.000 n=10)
ObjectGet/items_1000/lookups_4-8                       4.861µ ±  0%    5.335µ ±  0%   +9.76% (p=0.000 n=10)
ObjectGet/items_1000/lookups_8-8                       5.749µ ±  1%    6.279µ ±  0%   +9.21% (p=0.000 n=10)
ObjectGet/items_1000/lookups_16-8                      7.554µ ±  0%    8.134µ ±  0%   +7.68% (p=0.000 n=10)
ObjectGet/items_1000/lookups_32-8                      11.16µ ±  1%    11.88µ ±  0%   +6.54% (p=0.000 n=10)
ObjectGet/items_1000/lookups_64-8                      18.34µ ±  0%    19.33µ ±  1%   +5.40% (p=0.000 n=10)
ObjectGet/items_10000/lookups_0-8                      41.64µ ±  1%    46.44µ ±  0%  +11.52% (p=0.000 n=10)
ObjectGet/items_10000/lookups_1-8                      43.39µ ±  1%    48.20µ ±  1%  +11.08% (p=0.000 n=10)
ObjectGet/items_10000/lookups_2-8                      45.19µ ±  0%    50.04µ ±  0%  +10.74% (p=0.000 n=10)
ObjectGet/items_10000/lookups_4-8                      48.85µ ±  1%    53.67µ ±  0%   +9.86% (p=0.000 n=10)
ObjectGet/items_10000/lookups_8-8                      55.94µ ±  1%    61.00µ ±  0%   +9.04% (p=0.000 n=10)
ObjectGet/items_10000/lookups_16-8                     70.32µ ±  0%    75.42µ ±  0%   +7.26% (p=0.000 n=10)
ObjectGet/items_10000/lookups_32-8                     99.05µ ±  1%   104.36µ ±  0%   +5.36% (p=0.000 n=10)
ObjectGet/items_10000/lookups_64-8                     155.8µ ±  0%    161.9µ ±  1%   +3.91% (p=0.000 n=10)
ObjectGet/items_100000/lookups_0-8                     848.3µ ±  0%    786.0µ ± 14%   -7.34% (p=0.023 n=10)
ObjectGet/items_100000/lookups_1-8                     899.5µ ±  2%    842.3µ ±  1%   -6.36% (p=0.002 n=10)
ObjectGet/items_100000/lookups_2-8                     957.8µ ±  1%    898.8µ ± 13%   -6.16% (p=0.023 n=10)
ObjectGet/items_100000/lookups_4-8                    1038.3µ ±  1%    971.2µ ±  2%   -6.46% (p=0.000 n=10)
ObjectGet/items_100000/lookups_8-8                     1.188m ± 18%    1.088m ±  3%   -8.42% (p=0.000 n=10)
ObjectGet/items_100000/lookups_16-8                    1.408m ±  1%    1.312m ±  2%   -6.78% (p=0.000 n=10)
ObjectGet/items_100000/lookups_32-8                    2.036m ±  2%    1.913m ±  2%   -6.01% (p=0.000 n=10)
ObjectGet/items_100000/lookups_64-8                    3.189m ±  3%    3.031m ±  3%   -4.95% (p=0.000 n=10)
geomean                                                7.870µ          8.352µ         +6.13%

                                    │ BenchmarkObjectGet-before.txt │     BenchmarkObjectGet-after.txt      │
                                    │              B/s              │      B/s       vs base                │
ObjectGet/items_10/lookups_0-8                        5.254Gi ±  3%   4.449Gi ±  3%  -15.32% (p=0.000 n=10)
ObjectGet/items_10/lookups_1-8                        3.741Gi ±  2%   3.238Gi ±  3%  -13.42% (p=0.000 n=10)
ObjectGet/items_10/lookups_2-8                        3.389Gi ±  2%   2.965Gi ±  0%  -12.51% (p=0.000 n=10)
ObjectGet/items_10/lookups_4-8                        2.840Gi ±  2%   2.530Gi ±  2%  -10.91% (p=0.000 n=10)
ObjectGet/items_10/lookups_8-8                        2.218Gi ±  2%   2.037Gi ±  2%   -8.19% (p=0.000 n=10)
ObjectGet/items_10/lookups_16-8                       1.536Gi ±  1%   1.424Gi ±  1%   -7.29% (p=0.000 n=10)
ObjectGet/items_10/lookups_32-8                       976.1Mi ±  0%   921.5Mi ±  1%   -5.59% (p=0.000 n=10)
ObjectGet/items_10/lookups_64-8                       554.7Mi ±  1%   527.3Mi ±  0%   -4.93% (p=0.000 n=10)
ObjectGet/items_100/lookups_0-8                       5.040Gi ±  0%   4.519Gi ±  1%  -10.34% (p=0.000 n=10)
ObjectGet/items_100/lookups_1-8                       4.707Gi ±  1%   4.212Gi ±  1%  -10.51% (p=0.000 n=10)
ObjectGet/items_100/lookups_2-8                       4.426Gi ±  0%   3.964Gi ±  0%  -10.44% (p=0.000 n=10)
ObjectGet/items_100/lookups_4-8                       3.965Gi ±  0%   3.565Gi ±  0%  -10.10% (p=0.000 n=10)
ObjectGet/items_100/lookups_8-8                       3.296Gi ±  1%   2.996Gi ±  0%   -9.09% (p=0.000 n=10)
ObjectGet/items_100/lookups_16-8                      2.465Gi ±  0%   2.261Gi ±  0%   -8.26% (p=0.000 n=10)
ObjectGet/items_100/lookups_32-8                      1.635Gi ±  1%   1.525Gi ±  0%   -6.75% (p=0.000 n=10)
ObjectGet/items_100/lookups_64-8                     1004.6Mi ±  0%   937.9Mi ±  0%   -6.63% (p=0.000 n=10)
ObjectGet/items_1000/lookups_0-8                      5.375Gi ±  0%   4.828Gi ±  0%  -10.17% (p=0.000 n=10)
ObjectGet/items_1000/lookups_1-8                      5.078Gi ±  0%   4.594Gi ±  0%   -9.53% (p=0.000 n=10)
ObjectGet/items_1000/lookups_2-8                      4.821Gi ±  0%   4.362Gi ±  1%   -9.52% (p=0.000 n=10)
ObjectGet/items_1000/lookups_4-8                      4.365Gi ±  0%   3.977Gi ±  0%   -8.89% (p=0.000 n=10)
ObjectGet/items_1000/lookups_8-8                      3.691Gi ±  1%   3.379Gi ±  0%   -8.43% (p=0.000 n=10)
ObjectGet/items_1000/lookups_16-8                     2.809Gi ±  0%   2.608Gi ±  0%   -7.13% (p=0.000 n=10)
ObjectGet/items_1000/lookups_32-8                     1.902Gi ±  1%   1.785Gi ±  0%   -6.14% (p=0.000 n=10)
ObjectGet/items_1000/lookups_64-8                     1.157Gi ±  0%   1.097Gi ±  1%   -5.13% (p=0.000 n=10)
ObjectGet/items_10000/lookups_0-8                     5.541Gi ±  1%   4.969Gi ±  0%  -10.33% (p=0.000 n=10)
ObjectGet/items_10000/lookups_1-8                     5.318Gi ±  1%   4.787Gi ±  1%   -9.97% (p=0.000 n=10)
ObjectGet/items_10000/lookups_2-8                     5.107Gi ±  0%   4.612Gi ±  0%   -9.70% (p=0.000 n=10)
ObjectGet/items_10000/lookups_4-8                     4.724Gi ±  1%   4.300Gi ±  0%   -8.98% (p=0.000 n=10)
ObjectGet/items_10000/lookups_8-8                     4.125Gi ±  1%   3.783Gi ±  0%   -8.29% (p=0.000 n=10)
ObjectGet/items_10000/lookups_16-8                    3.282Gi ±  0%   3.060Gi ±  0%   -6.77% (p=0.000 n=10)
ObjectGet/items_10000/lookups_32-8                    2.330Gi ±  1%   2.211Gi ±  0%   -5.08% (p=0.000 n=10)
ObjectGet/items_10000/lookups_64-8                    1.481Gi ±  0%   1.426Gi ±  1%   -3.76% (p=0.000 n=10)
ObjectGet/items_100000/lookups_0-8                    2.940Gi ±  0%   3.173Gi ± 13%   +7.92% (p=0.023 n=10)
ObjectGet/items_100000/lookups_1-8                    2.773Gi ±  2%   2.961Gi ±  1%   +6.78% (p=0.002 n=10)
ObjectGet/items_100000/lookups_2-8                    2.604Gi ±  1%   2.775Gi ± 12%   +6.57% (p=0.023 n=10)
ObjectGet/items_100000/lookups_4-8                    2.402Gi ±  1%   2.568Gi ±  2%   +6.91% (p=0.000 n=10)
ObjectGet/items_100000/lookups_8-8                    2.100Gi ± 16%   2.293Gi ±  3%   +9.19% (p=0.000 n=10)
ObjectGet/items_100000/lookups_16-8                   1.772Gi ±  1%   1.900Gi ±  2%   +7.27% (p=0.000 n=10)
ObjectGet/items_100000/lookups_32-8                   1.225Gi ±  2%   1.303Gi ±  2%   +6.39% (p=0.000 n=10)
ObjectGet/items_100000/lookups_64-8                   800.8Mi ±  3%   842.5Mi ±  3%   +5.21% (p=0.000 n=10)
geomean                                               2.684Gi         2.529Gi         -5.77%

                                    │ BenchmarkObjectGet-before.txt │      BenchmarkObjectGet-after.txt       │
                                    │             B/op              │     B/op       vs base                  │
ObjectGet/items_10/lookups_0-8                        0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
ObjectGet/items_10/lookups_1-8                        0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
ObjectGet/items_10/lookups_2-8                        0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
ObjectGet/items_10/lookups_4-8                        0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
ObjectGet/items_10/lookups_8-8                        0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=10) ¹
…